Refactor decorator methods
