+++ backend/src/decorators/UseLoggers.ts
@@ -1,16 +1,11 @@
 import { ExcecutionLogger } from 'src/loggers/ExecutionLogger';
-import {
-  createClassExecutorDecorator,
-  DescriptorMethodDecorator,
-  methodExecutorDecorator,
-  TargetClassDecorator,
-} from 'src/utils/decorators';
-import { Class, InstanciableArray } from 'src/utils/types';
+import { Class, DescriptorMethodDecorator, InstanciableArray, TargetClassDecorator } from 'src/utils/types';
+import { ExecutorDecoratorFactory } from './ExecutorDecoratorFactory';
 
 export function UseLoggers(...loggers: Array<ExcecutionLogger>): TargetClassDecorator;
 export function UseLoggers(...loggers: Array<Class<ExcecutionLogger>>): TargetClassDecorator;
 export function UseLoggers(...loggers: InstanciableArray<ExcecutionLogger>): TargetClassDecorator {
-  return createClassExecutorDecorator(MethodUseLoggers, ...loggers);
+  return ExecutorDecoratorFactory.classDecorator(MethodUseLoggers, ...loggers);
 }
 
 export function MethodUseLoggers(...loggers: Array<ExcecutionLogger>): DescriptorMethodDecorator;
@@ -18,5 +13,5 @@ export function MethodUseLoggers(...loggers: Array<Class<ExcecutionLogger>>): De
 export function MethodUseLoggers(
   ...loggers: InstanciableArray<ExcecutionLogger>
 ): DescriptorMethodDecorator {
-  return methodExecutorDecorator(...loggers);
+  return ExecutorDecoratorFactory.methodDecorator(...loggers);
 }
